requirements.txt
einops
matplotlib
numpy
pandas
PyYAML
scikit_learn
scipy
sktime
torchmetrics
wget
pytorch_lightning==1.9.0
wandb